✏️ Componentize
